requirements.txt
django-ckeditor==6.7.1
Django==5.0.2
